% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_precision.R
\name{print.critical_precision}
\alias{print.critical_precision}
\title{Print method for critical_precision objects}
\usage{
\method{print}{critical_precision}(x, ...)
}
\arguments{
\item{x}{A critical_precision object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for critical_precision objects
}
