% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_design.R
\name{print.required_design}
\alias{print.required_design}
\title{Print method for required_design objects}
\usage{
\method{print}{required_design}(x, ...)
}
\arguments{
\item{x}{A required_design object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for required_design objects
}
